+++ modinfo.lua
@@ -143,6 +143,22 @@ configuration_options = {
         default = "KEY_CTRL + KEY_Z",
         hover = "Press this button to undo last cancel.",
     },
+    {
+        name = "keyToRepeatLastCraft",
+        label = "Repeat last craft",
+        options = 
+        {
+            { description = "None", data = false   },
+            { description = "Z",    data = "KEY_Z" },
+            { description = "X",    data = "KEY_X" },
+            { description = "C",    data = "KEY_C" },
+            { description = "V",    data = "KEY_V" },
+            { description = "B",    data = "KEY_B" },
+            { description = "W",    data = "KEY_W" },
+        },
+        default = false,
+        hover = "Press this button to repeat last craft.",
+    },
 
     -- Double click