+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Tweaks and features:
 * Add option to use W key to support azerty layout
+* Add option to repeat last craft via a configurable hotkey
 
 Fixes:
 * Fix ongoing search from double click conflicting with deselect feature